third_party/blink/web_tests/http/tests/misc/resource-timing-iframe-restored-from-history-cross-origin-expected.txt

ALERT: Initial URL loaded.
ALERT: Final URL loaded.
ALERT: Going back.
ALERT: Final URL loaded.
Tests that iframe restored from history does not report resource timing without explicitly setting src.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS resources.length is 2
PASS resources[0].name is "http://127.0.0.1:8000/js-test-resources/js-test.js"
PASS resources[1].name is "http://127.0.0.1:8000/resources/prevent-bfcache.js"
PASS successfullyParsed is true

TEST COMPLETE


Back-forward cache keeps and restores the resource timing of the iframe on back navigation. Back-forward cache needs to be disabled to run this test.
